Changed Data Storage Method
Changed Data Storage method from a SQL-Lite environment to a .JSON storage type
Rather than saving files to a Database, saving the notes and "events" to a .JSON file allows for easy parsing and hierachial structure for the program to save data.

This will also make it easy for users to port data out of the application for outside use.